+++ src/Game/Scenes/Settings/SettingsScene.cpp
@@ -27,7 +27,7 @@ namespace RType
 
                 int h = _linkSpacing;
                 for (auto &link : _links) {
-                    std::get<1>(link)->setPosition({(int)(_linkSpacing * 1.5), h});
+                    std::get<1>(link)->setPosition({(float)(_linkSpacing * 1.5), (float)h});
                     h += _linkGap + std::get<1>(link)->getSize().y;
                     std::get<2>(link)->load({(float)(window.getWidth() / 5 * 2 + _linkSpacing), (float)(_linkSpacing)}, {(float)(window.getWidth() / 5 * 3 - _linkSpacing * 2), (float)(window.getHeight() - _linkSpacing * 2)});
                 }
@@ -56,7 +56,7 @@ namespace RType
                 int h = _linkSpacing;
                 for (auto &link : _links) {
                     std::get<1>(link)->setText(Traductor::get()->translate(std::get<0>(link)));
-                    std::get<1>(link)->setPosition({(int)(_linkSpacing * 1.5), h});
+                    std::get<1>(link)->setPosition({(float)(_linkSpacing * 1.5), (float)h});
                     h += _linkGap + std::get<1>(link)->getSize().y;
                     std::get<2>(link)->reload({(float)(window.getWidth() / 5 * 2 + _linkSpacing), (float)(_linkSpacing)}, {(float)(window.getWidth() / 5 * 3 - _linkSpacing * 2), (float)(window.getHeight() - _linkSpacing * 2)});
                 }
